requirements.txt
pandas
openpyxl
matplotlib
openmeteo-requests 
requests-cache 
retry-requests
scikit-learn
holidays
seaborn
xgboost 
lightgbm 
catboost
statsmodels
optuna
streamlit
python-dotenv
mlflow
fastapi
uvicorn
pydantic
dagshub